refactor pane parts
