+++ src/vs/workbench/browser/parts/activitybar/activitybarPart.ts
@@ -8,79 +8,43 @@ import { localize } from 'vs/nls';
 import { ActionsOrientation, ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
 import { GLOBAL_ACTIVITY_ID, IActivity, ACCOUNTS_ACTIVITY_ID } from 'vs/workbench/common/activity';
 import { Part } from 'vs/workbench/browser/part';
-import { GlobalActivityActionViewItem, ViewContainerActivityAction, PlaceHolderToggleCompositePinnedAction, PlaceHolderViewContainerActivityAction, AccountsActivityActionViewItem, PlaceHolderToggleCompositeBadgeAction } from 'vs/workbench/browser/parts/activitybar/activitybarActions';
+import { GlobalActivityActionViewItem, AccountsActivityActionViewItem } from 'vs/workbench/browser/parts/activitybar/activitybarActions';
 import { IBadge, NumberBadge } from 'vs/workbench/services/activity/common/activity';
 import { IWorkbenchLayoutService, Parts, Position } from 'vs/workbench/services/layout/browser/layoutService';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { IDisposable, toDisposable, DisposableStore, Disposable } from 'vs/base/common/lifecycle';
+import { IDisposable, toDisposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { ToggleActivityBarVisibilityAction, ToggleSidebarPositionAction } from 'vs/workbench/browser/actions/layoutActions';
 import { IThemeService, IColorTheme } from 'vs/platform/theme/common/themeService';
 import { ACTIVITY_BAR_BACKGROUND, ACTIVITY_BAR_BORDER, ACTIVITY_BAR_FOREGROUND, ACTIVITY_BAR_ACTIVE_BORDER, ACTIVITY_BAR_BADGE_BACKGROUND, ACTIVITY_BAR_BADGE_FOREGROUND, ACTIVITY_BAR_INACTIVE_FOREGROUND, ACTIVITY_BAR_ACTIVE_BACKGROUND, ACTIVITY_BAR_DRAG_AND_DROP_BORDER } from 'vs/workbench/common/theme';
 import { contrastBorder } from 'vs/platform/theme/common/colorRegistry';
-import { CompositeBar, ICompositeBarItem, CompositeDragAndDrop } from 'vs/workbench/browser/parts/compositeBar';
-import { Dimension, createCSSRule, asCSSUrl, addDisposableListener, EventType, isAncestor } from 'vs/base/browser/dom';
-import { IStorageService, StorageScope, StorageTarget, IProfileStorageValueChangeEvent } from 'vs/platform/storage/common/storage';
+import { addDisposableListener, EventType, isAncestor } from 'vs/base/browser/dom';
+import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
-import { URI, UriComponents } from 'vs/base/common/uri';
-import { ToggleCompositePinnedAction, ICompositeBarColors, ActivityAction, ICompositeActivity, IActivityHoverOptions, ToggleCompositeBadgeAction } from 'vs/workbench/browser/parts/compositeBarActions';
-import { IViewDescriptorService, ViewContainer, IViewContainerModel, ViewContainerLocation } from 'vs/workbench/common/views';
-import { getEnabledViewContainerContextKey } from 'vs/workbench/common/contextkeys';
-import { IContextKeyService, ContextKeyExpr, IContextKey } from 'vs/platform/contextkey/common/contextkey';
-import { assertIsDefined, isString } from 'vs/base/common/types';
-import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { ICompositeBarColors, ActivityAction, ICompositeActivity, IActivityHoverOptions } from 'vs/workbench/browser/parts/compositeBarActions';
+import { assertIsDefined } from 'vs/base/common/types';
 import { CustomMenubarControl } from 'vs/workbench/browser/parts/titlebar/menubarControl';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { getMenuBarVisibility } from 'vs/platform/window/common/window';
-import { isNative } from 'vs/base/common/platform';
-import { Before2D } from 'vs/workbench/browser/dnd';
 import { Codicon } from 'vs/base/common/codicons';
 import { ThemeIcon } from 'vs/base/common/themables';
-import { IAction, Separator, toAction } from 'vs/base/common/actions';
+import { Separator, toAction } from 'vs/base/common/actions';
 import { StandardKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { KeyCode } from 'vs/base/common/keyCodes';
 import { registerIcon } from 'vs/platform/theme/common/iconRegistry';
-import { StringSHA1 } from 'vs/base/common/hash';
 import { HoverPosition } from 'vs/base/browser/ui/hover/hoverWidget';
 import { GestureEvent } from 'vs/base/browser/touch';
 import { IPaneCompositePart, IPaneCompositeSelectorPart } from 'vs/workbench/browser/parts/paneCompositePart';
 import { IUserDataProfileService } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
 import { DEFAULT_ICON } from 'vs/workbench/services/userDataProfile/common/userDataProfileIcons';
-
-interface IPlaceholderViewContainer {
-	readonly id: string;
-	readonly name?: string;
-	readonly iconUrl?: UriComponents;
-	readonly themeIcon?: ThemeIcon;
-	readonly isBuiltin?: boolean;
-	readonly views?: { when?: string }[];
-}
-
-interface IPinnedViewContainer {
-	readonly id: string;
-	readonly pinned: boolean;
-	readonly order?: number;
-	readonly visible: boolean;
-}
-
-interface ICachedViewContainer {
-	readonly id: string;
-	name?: string;
-	icon?: URI | ThemeIcon;
-	readonly pinned: boolean;
-	readonly order?: number;
-	visible: boolean;
-	isBuiltin?: boolean;
-	views?: { when?: string }[];
-}
+import { PaneCompositeBar } from 'vs/workbench/browser/parts/paneCompositeBar';
 
 export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart {
 
-	declare readonly _serviceBrand: undefined;
-
-	private static readonly PINNED_VIEW_CONTAINERS = 'workbench.activity.pinnedViewlets2';
-	private static readonly PLACEHOLDER_VIEW_CONTAINERS = 'workbench.activity.placeholderViewlets';
 	private static readonly ACTION_HEIGHT = 48;
 	private static readonly ACCOUNTS_ACTION_INDEX = 0;
+
+	static readonly pinnedViewContainersKey = 'workbench.activity.pinnedViewlets2';
+	static readonly placeholderViewContainersKey = 'workbench.activity.placeholderViewlets';
 
 	private static readonly ACCOUNTS_ICON = registerIcon('accounts-view-bar-icon', Codicon.account, localize('accountsViewBarIcon', "Accounts icon in the view bar."));
 
@@ -98,8 +62,8 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 	private menuBar: CustomMenubarControl | undefined;
 	private menuBarContainer: HTMLElement | undefined;
 
-	private compositeBar: CompositeBar;
 	private compositeBarContainer: HTMLElement | undefined;
+	private readonly compositeBar: PaneCompositeBar;
 
 	private globalActivityAction: ActivityAction | undefined;
 	private globalActivityActionBar: ActionBar | undefined;
@@ -107,18 +71,9 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 	private readonly globalActivity: ICompositeActivity[] = [];
 
 	private accountsActivityAction: ActivityAction | undefined;
-
 	private readonly accountsActivity: ICompositeActivity[] = [];
 
-	private readonly compositeActions = new Map<string, { activityAction: ViewContainerActivityAction; pinnedAction: ToggleCompositePinnedAction; badgeAction: ToggleCompositeBadgeAction }>();
-	private readonly viewContainerDisposables = new Map<string, IDisposable>();
-
 	private readonly keyboardNavigationDisposables = this._register(new DisposableStore());
-
-	private readonly location = ViewContainerLocation.Sidebar;
-	private hasExtensionsRegistered: boolean = false;
-
-	private readonly enabledViewContainersContextKeys: Map<string, IContextKey<boolean>> = new Map<string, IContextKey<boolean>>();
 
 	constructor(
 		private readonly paneCompositePart: IPaneCompositePart,
@@ -127,47 +82,24 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 		@IThemeService themeService: IThemeService,
 		@IStorageService private readonly storageService: IStorageService,
 		@IExtensionService private readonly extensionService: IExtensionService,
-		@IViewDescriptorService private readonly viewDescriptorService: IViewDescriptorService,
-		@IContextKeyService private readonly contextKeyService: IContextKeyService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
-		@IWorkbenchEnvironmentService private readonly environmentService: IWorkbenchEnvironmentService,
 		@IUserDataProfileService private readonly userDataProfileService: IUserDataProfileService,
 	) {
 		super(Parts.ACTIVITYBAR_PART, { hasTitle: false }, themeService, storageService, layoutService);
-
-		for (const cachedViewContainer of this.cachedViewContainers) {
-			cachedViewContainer.visible = !this.shouldBeHidden(cachedViewContainer.id, cachedViewContainer);
-		}
 		this.compositeBar = this.createCompositeBar();
-
-		this.onDidRegisterViewContainers(this.getViewContainers());
-
 		this.registerListeners();
-
 	}
 
-	private createCompositeBar() {
-		const cachedItems = this.cachedViewContainers
-			.map(container => ({
-				id: container.id,
-				name: container.name,
-				visible: container.visible,
-				order: container.order,
-				pinned: container.pinned,
-			}));
-
-		return this._register(this.instantiationService.createInstance(CompositeBar, cachedItems, {
-			icon: true,
+	private createCompositeBar(): PaneCompositeBar {
+		return this._register(this.instantiationService.createInstance(PaneCompositeBar, {
+			partContainerClass: 'activitybar',
+			pinnedViewContainersKey: ActivitybarPart.pinnedViewContainersKey,
+			placeholderViewContainersKey: ActivitybarPart.placeholderViewContainersKey,
 			orientation: ActionsOrientation.VERTICAL,
+			icon: true,
 			activityHoverOptions: this.getActivityHoverOptions(),
 			preventLoopNavigation: true,
-			openComposite: async (compositeId, preserveFocus) => {
-				return (await this.paneCompositePart.openPaneComposite(compositeId, !preserveFocus)) ?? null;
-			},
-			getActivityAction: compositeId => this.getCompositeActions(compositeId).activityAction,
-			getCompositePinnedAction: compositeId => this.getCompositeActions(compositeId).pinnedAction,
-			getCompositeBadgeAction: compositeId => this.getCompositeActions(compositeId).badgeAction,
-			getOnCompositeClickAction: compositeId => toAction({ id: compositeId, label: '', run: async () => this.paneCompositePart.getActivePaneComposite()?.getId() === compositeId ? this.paneCompositePart.hideActivePaneComposite() : this.paneCompositePart.openPaneComposite(compositeId) }),
+			recomputeSizes: false,
 			fillExtraContextMenuActions: (actions, e?: MouseEvent | GestureEvent) => {
 				// Menu
 				const menuBarVisibility = getMenuBarVisibility(this.configurationService);
@@ -192,18 +124,10 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 				// Toggle Activity Bar
 				actions.push(toAction({ id: ToggleActivityBarVisibilityAction.ID, label: localize('hideActivitBar', "Hide Activity Bar"), run: () => this.instantiationService.invokeFunction(accessor => new ToggleActivityBarVisibilityAction().run(accessor)) }));
 			},
-			getContextMenuActionsForComposite: compositeId => this.getContextMenuActionsForComposite(compositeId),
-			getDefaultCompositeId: () => this.viewDescriptorService.getDefaultViewContainer(this.location)?.id,
-			hidePart: () => this.layoutService.setPartHidden(true, Parts.SIDEBAR_PART),
-			dndHandler: new CompositeDragAndDrop(this.viewDescriptorService, ViewContainerLocation.Sidebar,
-				async (id: string, focus?: boolean) => { return await this.paneCompositePart.openPaneComposite(id, focus) ?? null; },
-				(from: string, to: string, before?: Before2D) => this.compositeBar.move(from, to, before?.verticallyBefore),
-				() => this.compositeBar.getCompositeBarItems(),
-			),
 			compositeSize: 52,
 			colors: (theme: IColorTheme) => this.getActivitybarItemColors(theme),
-			overflowActionSize: ActivitybarPart.ACTION_HEIGHT
-		}));
+			overflowActionSize: ActivitybarPart.ACTION_HEIGHT,
+		}, Parts.ACTIVITYBAR_PART, this.paneCompositePart));
 	}
 
 	private getActivityHoverOptions(): IActivityHoverOptions {
@@ -212,44 +136,11 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 		};
 	}
 
-	private getContextMenuActionsForComposite(compositeId: string): IAction[] {
-		const actions: IAction[] = [];
-
-		const viewContainer = this.viewDescriptorService.getViewContainerById(compositeId)!;
-		const defaultLocation = this.viewDescriptorService.getDefaultViewContainerLocation(viewContainer)!;
-		if (defaultLocation !== this.viewDescriptorService.getViewContainerLocation(viewContainer)) {
-			actions.push(toAction({ id: 'resetLocationAction', label: localize('resetLocation', "Reset Location"), run: () => this.viewDescriptorService.moveViewContainerToLocation(viewContainer, defaultLocation) }));
-		} else {
-			const viewContainerModel = this.viewDescriptorService.getViewContainerModel(viewContainer);
-			if (viewContainerModel.allViewDescriptors.length === 1) {
-				const viewToReset = viewContainerModel.allViewDescriptors[0];
-				const defaultContainer = this.viewDescriptorService.getDefaultContainerById(viewToReset.id)!;
-				if (defaultContainer !== viewContainer) {
-					actions.push(toAction({ id: 'resetLocationAction', label: localize('resetLocation', "Reset Location"), run: () => this.viewDescriptorService.moveViewsToContainer([viewToReset], defaultContainer) }));
-				}
-			}
-		}
-
-		return actions;
-	}
-
 	private registerListeners(): void {
-
-		// View Container Changes
-		this._register(this.viewDescriptorService.onDidChangeViewContainers(({ added, removed }) => this.onDidChangeViewContainers(added, removed)));
-		this._register(this.viewDescriptorService.onDidChangeContainerLocation(({ viewContainer, from, to }) => this.onDidChangeViewContainerLocation(viewContainer, from, to)));
-
-		// View Container Visibility Changes
-		this.paneCompositePart.onDidPaneCompositeOpen(e => this.onDidChangeViewContainerVisibility(e.getId(), true));
-		this.paneCompositePart.onDidPaneCompositeClose(e => this.onDidChangeViewContainerVisibility(e.getId(), false));
 
 		// Extension registration
 		const disposables = this._register(new DisposableStore());
 		this._register(this.extensionService.onDidRegisterExtensions(() => {
-			disposables.clear();
-			this.onDidRegisterExtensions();
-			this.compositeBar.onDidChange(() => this.saveCachedViewContainers(), this, disposables);
-			this.storageService.onDidChangeValue(StorageScope.PROFILE, ActivitybarPart.PINNED_VIEW_CONTAINERS, disposables)(e => this.onDidPinnedViewContainersStorageValueChange(e), this, disposables);
 			this.storageService.onDidChangeValue(StorageScope.PROFILE, AccountsActivityActionViewItem.ACCOUNTS_VISIBILITY_PREFERENCE_KEY, disposables)(() => this.toggleAccountsActivity(), this, disposables);
 		}));
 
@@ -265,74 +156,7 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 		}));
 	}
 
-	private onDidChangeViewContainers(added: readonly { container: ViewContainer; location: ViewContainerLocation }[], removed: readonly { container: ViewContainer; location: ViewContainerLocation }[]) {
-		removed.filter(({ location }) => location === ViewContainerLocation.Sidebar).forEach(({ container }) => this.onDidDeregisterViewContainer(container));
-		this.onDidRegisterViewContainers(added.filter(({ location }) => location === ViewContainerLocation.Sidebar).map(({ container }) => container));
-	}
-
-	private onDidChangeViewContainerLocation(container: ViewContainer, from: ViewContainerLocation, to: ViewContainerLocation) {
-		if (from === this.location) {
-			this.onDidDeregisterViewContainer(container);
-		}
-
-		if (to === this.location) {
-			this.onDidRegisterViewContainers([container]);
-		}
-	}
-
-	private onDidChangeViewContainerVisibility(id: string, visible: boolean) {
-		if (visible) {
-			// Activate view container action on opening of a view container
-			this.onDidViewContainerVisible(id);
-		} else {
-			// Deactivate view container action on close
-			this.compositeBar.deactivateComposite(id);
-		}
-	}
-
-	private onDidRegisterExtensions(): void {
-		this.hasExtensionsRegistered = true;
-
-		// show/hide/remove composites
-		for (const { id } of this.cachedViewContainers) {
-			const viewContainer = this.getViewContainer(id);
-			if (viewContainer) {
-				this.showOrHideViewContainer(viewContainer);
-			} else {
-				if (this.viewDescriptorService.isViewContainerRemovedPermanently(id)) {
-					this.removeComposite(id);
-				} else {
-					this.hideComposite(id);
-				}
-			}
-		}
-
-		this.saveCachedViewContainers();
-	}
-
-	private onDidViewContainerVisible(id: string): void {
-		const viewContainer = this.getViewContainer(id);
-		if (viewContainer) {
-
-			// Update the composite bar by adding
-			this.addComposite(viewContainer);
-			this.compositeBar.activateComposite(viewContainer.id);
-
-			if (this.shouldBeHidden(viewContainer)) {
-				const viewContainerModel = this.viewDescriptorService.getViewContainerModel(viewContainer);
-				if (viewContainerModel.activeViewDescriptors.length === 0) {
-					// Update the composite bar by hiding
-					this.hideComposite(viewContainer.id);
-				}
-			}
-		}
-	}
-
 	showActivity(viewContainerOrActionId: string, badge: IBadge, clazz?: string, priority?: number): IDisposable {
-		if (this.getViewContainer(viewContainerOrActionId)) {
-			return this.compositeBar.showActivity(viewContainerOrActionId, badge, clazz, priority);
-		}
-
 		if (viewContainerOrActionId === GLOBAL_ACTIVITY_ID) {
 			return this.showGlobalActivity(GLOBAL_ACTIVITY_ID, badge, clazz, priority);
 		}
@@ -341,7 +165,7 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 			return this.showGlobalActivity(ACCOUNTS_ACTIVITY_ID, badge, clazz, priority);
 		}
 
-		return Disposable.None;
+		return this.compositeBar.showActivity(viewContainerOrActionId, badge, clazz, priority);
 	}
 
 	private showGlobalActivity(activityId: string, badge: IBadge, clazz?: string, priority?: number): IDisposable {
@@ -574,192 +398,12 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 		this.updateGlobalActivity(ACCOUNTS_ACTIVITY_ID);
 	}
 
-	private getCompositeActions(compositeId: string): { activityAction: ViewContainerActivityAction; pinnedAction: ToggleCompositePinnedAction; badgeAction: ToggleCompositeBadgeAction } {
-		let compositeActions = this.compositeActions.get(compositeId);
-		if (!compositeActions) {
-			const viewContainer = this.getViewContainer(compositeId);
-			if (viewContainer) {
-				const viewContainerModel = this.viewDescriptorService.getViewContainerModel(viewContainer);
-				compositeActions = {
-					activityAction: this.instantiationService.createInstance(ViewContainerActivityAction, this.toActivity(viewContainerModel), this.paneCompositePart),
-					pinnedAction: new ToggleCompositePinnedAction(this.toActivity(viewContainerModel), this.compositeBar),
-					badgeAction: new ToggleCompositeBadgeAction(this.toActivity(viewContainerModel), this.compositeBar)
-				};
-			} else {
-				const cachedComposite = this.cachedViewContainers.filter(c => c.id === compositeId)[0];
-				compositeActions = {
-					activityAction: this.instantiationService.createInstance(PlaceHolderViewContainerActivityAction, ActivitybarPart.toActivity(compositeId, compositeId, cachedComposite?.icon, undefined), this.paneCompositePart),
-					pinnedAction: new PlaceHolderToggleCompositePinnedAction(compositeId, this.compositeBar),
-					badgeAction: new PlaceHolderToggleCompositeBadgeAction(compositeId, this.compositeBar)
-				};
-			}
-
-			this.compositeActions.set(compositeId, compositeActions);
-		}
-
-		return compositeActions;
-	}
-
-	private onDidRegisterViewContainers(viewContainers: readonly ViewContainer[]): void {
-		for (const viewContainer of viewContainers) {
-			this.addComposite(viewContainer);
-
-			// Pin it by default if it is new
-			const cachedViewContainer = this.cachedViewContainers.filter(({ id }) => id === viewContainer.id)[0];
-			if (!cachedViewContainer) {
-				this.compositeBar.pin(viewContainer.id);
-			}
-
-			// Active
-			const visibleViewContainer = this.paneCompositePart.getActivePaneComposite();
-			if (visibleViewContainer?.getId() === viewContainer.id) {
-				this.compositeBar.activateComposite(viewContainer.id);
-			}
-
-			const viewContainerModel = this.viewDescriptorService.getViewContainerModel(viewContainer);
-			this.updateActivity(viewContainer, viewContainerModel);
-			this.showOrHideViewContainer(viewContainer);
-
-			const disposables = new DisposableStore();
-			disposables.add(viewContainerModel.onDidChangeContainerInfo(() => this.updateActivity(viewContainer, viewContainerModel)));
-			disposables.add(viewContainerModel.onDidChangeActiveViewDescriptors(() => this.showOrHideViewContainer(viewContainer)));
-
-			this.viewContainerDisposables.set(viewContainer.id, disposables);
-		}
-	}
-
-	private onDidDeregisterViewContainer(viewContainer: ViewContainer): void {
-		const disposable = this.viewContainerDisposables.get(viewContainer.id);
-		disposable?.dispose();
-
-		this.viewContainerDisposables.delete(viewContainer.id);
-		this.removeComposite(viewContainer.id);
-	}
-
-	private updateActivity(viewContainer: ViewContainer, viewContainerModel: IViewContainerModel): void {
-		const activity: IActivity = this.toActivity(viewContainerModel);
-		const { activityAction, pinnedAction } = this.getCompositeActions(viewContainer.id);
-		activityAction.updateActivity(activity);
-
-		if (pinnedAction instanceof PlaceHolderToggleCompositePinnedAction) {
-			pinnedAction.setActivity(activity);
-		}
-
-		this.saveCachedViewContainers();
-	}
-
-	private toActivity(viewContainerModel: IViewContainerModel): IActivity {
-		return ActivitybarPart.toActivity(viewContainerModel.viewContainer.id, viewContainerModel.title, viewContainerModel.icon, viewContainerModel.keybindingId);
-	}
-
-	private static toActivity(id: string, name: string, icon: URI | ThemeIcon | undefined, keybindingId: string | undefined): IActivity {
-		let classNames: string[] | undefined = undefined;
-		let iconUrl: URI | undefined = undefined;
-		if (URI.isUri(icon)) {
-			iconUrl = icon;
-			const cssUrl = asCSSUrl(icon);
-			const hash = new StringSHA1();
-			hash.update(cssUrl);
-			const iconId = `activity-${id.replace(/\./g, '-')}-${hash.digest()}`;
-			const iconClass = `.monaco-workbench .activitybar .monaco-action-bar .action-label.${iconId}`;
-			classNames = [iconId, 'uri-icon'];
-			createCSSRule(iconClass, `
-				mask: ${cssUrl} no-repeat 50% 50%;
-				mask-size: 24px;
-				-webkit-mask: ${cssUrl} no-repeat 50% 50%;
-				-webkit-mask-size: 24px;
-			`);
-		} else if (ThemeIcon.isThemeIcon(icon)) {
-			classNames = ThemeIcon.asClassNameArray(icon);
-		}
-
-		return { id, name, classNames, iconUrl, keybindingId };
-	}
-
-	private showOrHideViewContainer(viewContainer: ViewContainer): void {
-		let contextKey = this.enabledViewContainersContextKeys.get(viewContainer.id);
-		if (!contextKey) {
-			contextKey = this.contextKeyService.createKey(getEnabledViewContainerContextKey(viewContainer.id), false);
-			this.enabledViewContainersContextKeys.set(viewContainer.id, contextKey);
-		}
-		if (this.shouldBeHidden(viewContainer)) {
-			contextKey.set(false);
-			this.hideComposite(viewContainer.id);
-		} else {
-			contextKey.set(true);
-			this.addComposite(viewContainer);
-		}
-	}
-
-	private shouldBeHidden(viewContainerOrId: string | ViewContainer, cachedViewContainer?: ICachedViewContainer): boolean {
-		const viewContainer = isString(viewContainerOrId) ? this.getViewContainer(viewContainerOrId) : viewContainerOrId;
-		const viewContainerId = isString(viewContainerOrId) ? viewContainerOrId : viewContainerOrId.id;
-
-		if (viewContainer) {
-			if (viewContainer.hideIfEmpty) {
-				if (this.viewDescriptorService.getViewContainerModel(viewContainer).activeViewDescriptors.length > 0) {
-					return false;
-				}
-			} else {
-				return false;
-			}
-		}
-
-		// Check cache only if extensions are not yet registered and current window is not native (desktop) remote connection window
-		if (!this.hasExtensionsRegistered && !(this.environmentService.remoteAuthority && isNative)) {
-			cachedViewContainer = cachedViewContainer || this.cachedViewContainers.find(({ id }) => id === viewContainerId);
-
-			// Show builtin ViewContainer if not registered yet
-			if (!viewContainer && cachedViewContainer?.isBuiltin && cachedViewContainer?.visible) {
-				return false;
-			}
-
-			if (cachedViewContainer?.views?.length) {
-				return cachedViewContainer.views.every(({ when }) => !!when && !this.contextKeyService.contextMatchesRules(ContextKeyExpr.deserialize(when)));
-			}
-		}
-
-		return true;
-	}
-
-	private addComposite(viewContainer: ViewContainer): void {
-		this.compositeBar.addComposite({ id: viewContainer.id, name: typeof viewContainer.title === 'string' ? viewContainer.title : viewContainer.title.value, order: viewContainer.order, requestedIndex: viewContainer.requestedIndex });
-	}
-
-	private hideComposite(compositeId: string): void {
-		this.compositeBar.hideComposite(compositeId);
-
-		const compositeActions = this.compositeActions.get(compositeId);
-		if (compositeActions) {
-			compositeActions.activityAction.dispose();
-			compositeActions.pinnedAction.dispose();
-			this.compositeActions.delete(compositeId);
-		}
-	}
-
-	private removeComposite(compositeId: string): void {
-		this.compositeBar.removeComposite(compositeId);
-
-		const compositeActions = this.compositeActions.get(compositeId);
-		if (compositeActions) {
-			compositeActions.activityAction.dispose();
-			compositeActions.pinnedAction.dispose();
-			this.compositeActions.delete(compositeId);
-		}
-	}
-
 	getPinnedPaneCompositeIds(): string[] {
-		const pinnedCompositeIds = this.compositeBar.getPinnedComposites().map(v => v.id);
-		return this.getViewContainers()
-			.filter(v => this.compositeBar.isPinned(v.id))
-			.sort((v1, v2) => pinnedCompositeIds.indexOf(v1.id) - pinnedCompositeIds.indexOf(v2.id))
-			.map(v => v.id);
+		return this.compositeBar.getPinnedPaneCompositeIds();
 	}
 
 	getVisiblePaneCompositeIds(): string[] {
-		return this.compositeBar.getVisibleComposites()
-			.filter(v => this.paneCompositePart.getActivePaneComposite()?.getId() === v.id || this.compositeBar.isPinned(v.id))
-			.map(v => v.id);
+		return this.compositeBar.getVisiblePaneCompositeIds();
 	}
 
 	focus(): void {
@@ -807,185 +451,7 @@ export class ActivitybarPart extends Part implements IPaneCompositeSelectorPart 
 		if (this.globalActivityActionBar) {
 			availableHeight -= (this.globalActivityActionBar.viewItems.length * ActivitybarPart.ACTION_HEIGHT); // adjust height for global actions showing
 		}
-		this.compositeBar.layout(new Dimension(width, availableHeight));
-	}
-
-	private getViewContainer(id: string): ViewContainer | undefined {
-		const viewContainer = this.viewDescriptorService.getViewContainerById(id);
-		return viewContainer && this.viewDescriptorService.getViewContainerLocation(viewContainer) === this.location ? viewContainer : undefined;
-	}
-
-	private getViewContainers(): readonly ViewContainer[] {
-		return this.viewDescriptorService.getViewContainersByLocation(this.location);
-	}
-
-	private onDidPinnedViewContainersStorageValueChange(e: IProfileStorageValueChangeEvent): void {
-		if (this.pinnedViewContainersValue !== this.getStoredPinnedViewContainersValue() /* This checks if current window changed the value or not */) {
-			this._pinnedViewContainersValue = undefined;
-			this._cachedViewContainers = undefined;
-
-			const newCompositeItems: ICompositeBarItem[] = [];
-			const compositeItems = this.compositeBar.getCompositeBarItems();
-
-			for (const cachedViewContainer of this.cachedViewContainers) {
-				newCompositeItems.push({
-					id: cachedViewContainer.id,
-					name: cachedViewContainer.name,
-					order: cachedViewContainer.order,
-					pinned: cachedViewContainer.pinned,
-					visible: !!compositeItems.find(({ id }) => id === cachedViewContainer.id)
-				});
-			}
-
-			for (let index = 0; index < compositeItems.length; index++) {
-				// Add items currently exists but does not exist in new.
-				if (!newCompositeItems.some(({ id }) => id === compositeItems[index].id)) {
-					const viewContainer = this.viewDescriptorService.getViewContainerById(compositeItems[index].id);
-					newCompositeItems.splice(index, 0, {
-						...compositeItems[index],
-						pinned: true,
-						visible: true,
-						order: viewContainer?.order,
-					});
-				}
-			}
-
-			this.compositeBar.setCompositeBarItems(newCompositeItems);
-		}
-	}
-
-	private saveCachedViewContainers(): void {
-		const state: ICachedViewContainer[] = [];
-
-		const compositeItems = this.compositeBar.getCompositeBarItems();
-		for (const compositeItem of compositeItems) {
-			const viewContainer = this.getViewContainer(compositeItem.id);
-			if (viewContainer) {
-				const viewContainerModel = this.viewDescriptorService.getViewContainerModel(viewContainer);
-				const views: { when: string | undefined }[] = [];
-				for (const { when } of viewContainerModel.allViewDescriptors) {
-					views.push({ when: when ? when.serialize() : undefined });
-				}
-				state.push({
-					id: compositeItem.id,
-					name: viewContainerModel.title,
-					icon: URI.isUri(viewContainerModel.icon) && this.environmentService.remoteAuthority ? undefined : viewContainerModel.icon, // Do not cache uri icons with remote connection
-					views,
-					pinned: compositeItem.pinned,
-					order: compositeItem.order,
-					visible: compositeItem.visible,
-					isBuiltin: !viewContainer.extensionId
-				});
-			} else {
-				state.push({ id: compositeItem.id, pinned: compositeItem.pinned, order: compositeItem.order, visible: false, isBuiltin: false });
-			}
-		}
-
-		this.storeCachedViewContainersState(state);
-	}
-
-	private _cachedViewContainers: ICachedViewContainer[] | undefined = undefined;
-	private get cachedViewContainers(): ICachedViewContainer[] {
-		if (this._cachedViewContainers === undefined) {
-			this._cachedViewContainers = this.getPinnedViewContainers();
-			for (const placeholderViewContainer of this.getPlaceholderViewContainers()) {
-				const cachedViewContainer = this._cachedViewContainers.filter(cached => cached.id === placeholderViewContainer.id)[0];
-				if (cachedViewContainer) {
-					cachedViewContainer.name = placeholderViewContainer.name;
-					cachedViewContainer.icon = placeholderViewContainer.themeIcon ? placeholderViewContainer.themeIcon :
-						placeholderViewContainer.iconUrl ? URI.revive(placeholderViewContainer.iconUrl) : undefined;
-					if (URI.isUri(cachedViewContainer.icon) && this.environmentService.remoteAuthority) {
-						cachedViewContainer.icon = undefined; // Do not cache uri icons with remote connection
-					}
-					cachedViewContainer.views = placeholderViewContainer.views;
-					cachedViewContainer.isBuiltin = placeholderViewContainer.isBuiltin;
-				}
-			}
-		}
-
-		return this._cachedViewContainers;
-	}
-
-	private storeCachedViewContainersState(cachedViewContainers: ICachedViewContainer[]): void {
-		this.setPinnedViewContainers(cachedViewContainers.map(({ id, pinned, visible, order }) => (<IPinnedViewContainer>{
-			id,
-			pinned,
-			visible,
-			order
-		})));
-
-		this.setPlaceholderViewContainers(cachedViewContainers.map(({ id, icon, name, views, isBuiltin }) => (<IPlaceholderViewContainer>{
-			id,
-			iconUrl: URI.isUri(icon) ? icon : undefined,
-			themeIcon: ThemeIcon.isThemeIcon(icon) ? icon : undefined,
-			name,
-			isBuiltin,
-			views
-		})));
-	}
-
-	private getPinnedViewContainers(): IPinnedViewContainer[] {
-		return JSON.parse(this.pinnedViewContainersValue);
-	}
-
-	private setPinnedViewContainers(pinnedViewContainers: IPinnedViewContainer[]): void {
-		this.pinnedViewContainersValue = JSON.stringify(pinnedViewContainers);
-	}
-
-	private _pinnedViewContainersValue: string | undefined;
-	private get pinnedViewContainersValue(): string {
-		if (!this._pinnedViewContainersValue) {
-			this._pinnedViewContainersValue = this.getStoredPinnedViewContainersValue();
-		}
-
-		return this._pinnedViewContainersValue;
-	}
-
-	private set pinnedViewContainersValue(pinnedViewContainersValue: string) {
-		if (this.pinnedViewContainersValue !== pinnedViewContainersValue) {
-			this._pinnedViewContainersValue = pinnedViewContainersValue;
-			this.setStoredPinnedViewContainersValue(pinnedViewContainersValue);
-		}
-	}
-
-	private getStoredPinnedViewContainersValue(): string {
-		return this.storageService.get(ActivitybarPart.PINNED_VIEW_CONTAINERS, StorageScope.PROFILE, '[]');
-	}
-
-	private setStoredPinnedViewContainersValue(value: string): void {
-		this.storageService.store(ActivitybarPart.PINNED_VIEW_CONTAINERS, value, StorageScope.PROFILE, StorageTarget.USER);
-	}
-
-	private getPlaceholderViewContainers(): IPlaceholderViewContainer[] {
-		return JSON.parse(this.placeholderViewContainersValue);
-	}
-
-	private setPlaceholderViewContainers(placeholderViewContainers: IPlaceholderViewContainer[]): void {
-		this.placeholderViewContainersValue = JSON.stringify(placeholderViewContainers);
-	}
-
-	private _placeholderViewContainersValue: string | undefined;
-	private get placeholderViewContainersValue(): string {
-		if (!this._placeholderViewContainersValue) {
-			this._placeholderViewContainersValue = this.getStoredPlaceholderViewContainersValue();
-		}
-
-		return this._placeholderViewContainersValue;
-	}
-
-	private set placeholderViewContainersValue(placeholderViewContainesValue: string) {
-		if (this.placeholderViewContainersValue !== placeholderViewContainesValue) {
-			this._placeholderViewContainersValue = placeholderViewContainesValue;
-			this.setStoredPlaceholderViewContainersValue(placeholderViewContainesValue);
-		}
-	}
-
-	private getStoredPlaceholderViewContainersValue(): string {
-		return this.storageService.get(ActivitybarPart.PLACEHOLDER_VIEW_CONTAINERS, StorageScope.PROFILE, '[]');
-	}
-
-	private setStoredPlaceholderViewContainersValue(value: string): void {
-		this.storageService.store(ActivitybarPart.PLACEHOLDER_VIEW_CONTAINERS, value, StorageScope.PROFILE, StorageTarget.MACHINE);
+		this.compositeBar.layout(width, availableHeight);
 	}
 
 	private get accountsVisibilityPreference(): boolean {
