+++ src/vs/workbench/browser/parts/auxiliarybar/auxiliaryBarPart.ts
@@ -11,16 +11,13 @@ import { IInstantiationService } from 'vs/platform/instantiation/common/instanti
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { INotificationService } from 'vs/platform/notification/common/notification';
 import { IStorageService } from 'vs/platform/storage/common/storage';
-import { contrastBorder } from 'vs/platform/theme/common/colorRegistry';
+import { badgeBackground, badgeForeground, contrastBorder } from 'vs/platform/theme/common/colorRegistry';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { Extensions as PaneCompositeExtensions } from 'vs/workbench/browser/panecomposite';
-import { BasePanelPart } from 'vs/workbench/browser/parts/panel/panelPart';
 import { ActiveAuxiliaryContext, AuxiliaryBarFocusContext } from 'vs/workbench/common/contextkeys';
-import { SIDE_BAR_BACKGROUND, SIDE_BAR_BORDER, SIDE_BAR_FOREGROUND } from 'vs/workbench/common/theme';
-import { IViewDescriptorService, ViewContainerLocation } from 'vs/workbench/common/views';
+import { PANEL_ACTIVE_TITLE_BORDER, PANEL_ACTIVE_TITLE_FOREGROUND, PANEL_DRAG_AND_DROP_BORDER, PANEL_INACTIVE_TITLE_FOREGROUND, SIDE_BAR_BACKGROUND, SIDE_BAR_BORDER, SIDE_BAR_FOREGROUND } from 'vs/workbench/common/theme';
+import { IViewDescriptorService } from 'vs/workbench/common/views';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IWorkbenchLayoutService, Parts, Position } from 'vs/workbench/services/layout/browser/layoutService';
-import { IActivityHoverOptions } from 'vs/workbench/browser/parts/compositeBarActions';
 import { HoverPosition } from 'vs/base/browser/ui/hover/hoverWidget';
 import { IAction, Separator, toAction } from 'vs/base/common/actions';
 import { ToggleAuxiliaryBarAction } from 'vs/workbench/browser/parts/auxiliarybar/auxiliaryBarActions';
@@ -28,8 +25,11 @@ import { assertIsDefined } from 'vs/base/common/types';
 import { LayoutPriority } from 'vs/base/browser/ui/splitview/splitview';
 import { ToggleSidebarPositionAction } from 'vs/workbench/browser/actions/layoutActions';
 import { ICommandService } from 'vs/platform/commands/common/commands';
+import { AbstractPaneCompositePart } from 'vs/workbench/browser/parts/paneCompositePart';
+import { ActionsOrientation } from 'vs/base/browser/ui/actionbar/actionbar';
 
-export class AuxiliaryBarPart extends BasePanelPart {
+export class AuxiliaryBarPart extends AbstractPaneCompositePart {
+
 	static readonly activePanelSettingsKey = 'workbench.auxiliarybar.activepanelid';
 	static readonly pinnedPanelsKey = 'workbench.auxiliarybar.pinnedPanels';
 	static readonly placeholdeViewContainersKey = 'workbench.auxiliarybar.placeholderPanels';
@@ -39,6 +39,27 @@ export class AuxiliaryBarPart extends BasePanelPart {
 	override readonly maximumWidth: number = Number.POSITIVE_INFINITY;
 	override readonly minimumHeight: number = 0;
 	override readonly maximumHeight: number = Number.POSITIVE_INFINITY;
+
+	get preferredHeight(): number | undefined {
+		// Don't worry about titlebar or statusbar visibility
+		// The difference is minimal and keeps this function clean
+		return this.layoutService.dimension.height * 0.4;
+	}
+
+	get preferredWidth(): number | undefined {
+		const activeComposite = this.getActivePaneComposite();
+
+		if (!activeComposite) {
+			return;
+		}
+
+		const width = activeComposite.getOptimalWidth();
+		if (typeof width !== 'number') {
+			return;
+		}
+
+		return Math.max(width, 300);
+	}
 
 	readonly priority: LayoutPriority = LayoutPriority.Low;
 
@@ -56,6 +77,41 @@ export class AuxiliaryBarPart extends BasePanelPart {
 		@ICommandService private commandService: ICommandService,
 	) {
 		super(
+			Parts.AUXILIARYBAR_PART,
+			{
+				hasTitle: true,
+				borderWidth: () => (this.getColor(SIDE_BAR_BORDER) || this.getColor(contrastBorder)) ? 1 : 0,
+			},
+			{
+				partContainerClass: 'auxiliarybar',
+				pinnedViewContainersKey: AuxiliaryBarPart.pinnedPanelsKey,
+				placeholderViewContainersKey: AuxiliaryBarPart.placeholdeViewContainersKey,
+				icon: true,
+				orientation: ActionsOrientation.HORIZONTAL,
+				recomputeSizes: true,
+				activityHoverOptions: {
+					position: () => this.getActivityHoverPosition(),
+				},
+				fillExtraContextMenuActions: actions => this.fillExtraContextMenuActions(actions),
+				compositeSize: 0,
+				overflowActionSize: 44,
+				colors: theme => ({
+					activeBackgroundColor: theme.getColor(SIDE_BAR_BACKGROUND),
+					inactiveBackgroundColor: theme.getColor(SIDE_BAR_BACKGROUND),
+					activeBorderBottomColor: theme.getColor(PANEL_ACTIVE_TITLE_BORDER),
+					activeForegroundColor: theme.getColor(PANEL_ACTIVE_TITLE_FOREGROUND),
+					inactiveForegroundColor: theme.getColor(PANEL_INACTIVE_TITLE_FOREGROUND),
+					badgeBackground: theme.getColor(badgeBackground),
+					badgeForeground: theme.getColor(badgeForeground),
+					dragAndDropBorder: theme.getColor(PANEL_DRAG_AND_DROP_BORDER)
+				})
+			},
+			AuxiliaryBarPart.activePanelSettingsKey,
+			ActiveAuxiliaryContext.bindTo(contextKeyService),
+			AuxiliaryBarFocusContext.bindTo(contextKeyService),
+			'auxiliarybar',
+			'auxiliarybar',
+			undefined,
 			notificationService,
 			storageService,
 			contextMenuService,
@@ -66,20 +122,6 @@ export class AuxiliaryBarPart extends BasePanelPart {
 			viewDescriptorService,
 			contextKeyService,
 			extensionService,
-			Parts.AUXILIARYBAR_PART,
-			AuxiliaryBarPart.activePanelSettingsKey,
-			AuxiliaryBarPart.pinnedPanelsKey,
-			AuxiliaryBarPart.placeholdeViewContainersKey,
-			PaneCompositeExtensions.Auxiliary,
-			SIDE_BAR_BACKGROUND,
-			ViewContainerLocation.AuxiliaryBar,
-			ActiveAuxiliaryContext.bindTo(contextKeyService),
-			AuxiliaryBarFocusContext.bindTo(contextKeyService),
-			{
-				useIcons: true,
-				hasTitle: true,
-				borderWidth: () => (this.getColor(SIDE_BAR_BORDER) || this.getColor(contrastBorder)) ? 1 : 0,
-			}
 		);
 	}
 
@@ -87,6 +129,7 @@ export class AuxiliaryBarPart extends BasePanelPart {
 		super.updateStyles();
 
 		const container = assertIsDefined(this.getContainer());
+		container.style.backgroundColor = this.getColor(SIDE_BAR_BACKGROUND) || '';
 		const borderColor = this.getColor(SIDE_BAR_BORDER) || this.getColor(contrastBorder);
 		const isPositionLeft = this.layoutService.getSideBarPosition() === Position.RIGHT;
 
@@ -102,19 +145,21 @@ export class AuxiliaryBarPart extends BasePanelPart {
 		container.style.borderRightWidth = borderColor && isPositionLeft ? '1px' : '0px';
 	}
 
-	protected getActivityHoverOptions(): IActivityHoverOptions {
-		return {
-			position: () => HoverPosition.BELOW
-		};
+	private getActivityHoverPosition(): HoverPosition {
+		return HoverPosition.BELOW;
 	}
 
-	protected fillExtraContextMenuActions(actions: IAction[]): void {
+	private fillExtraContextMenuActions(actions: IAction[]): void {
 		const currentPositionRight = this.layoutService.getSideBarPosition() === Position.LEFT;
 		actions.push(...[
 			new Separator(),
 			toAction({ id: ToggleSidebarPositionAction.ID, label: currentPositionRight ? localize('move second side bar left', "Move Secondary Side Bar Left") : localize('move second side bar right', "Move Secondary Side Bar Right"), run: () => this.commandService.executeCommand(ToggleSidebarPositionAction.ID) }),
 			toAction({ id: ToggleAuxiliaryBarAction.ID, label: localize('hide second side bar', "Hide Secondary Side Bar"), run: () => this.commandService.executeCommand(ToggleAuxiliaryBarAction.ID) })
 		]);
+	}
+
+	protected shouldShowCompositeBar(): boolean {
+		return true;
 	}
 
 	override toJSON(): object {
