+++ src/vs/workbench/browser/parts/auxiliarybar/media/auxiliaryBarPart.css
@@ -14,16 +14,16 @@
 	background-color: var(--vscode-sideBar-background);
 }
 
-.monaco-workbench .part.auxiliarybar > .title > .panel-switcher-container > .monaco-action-bar .action-item:hover .action-label,
-.monaco-workbench .part.auxiliarybar > .title > .panel-switcher-container > .monaco-action-bar .action-item:focus .action-label {
+.monaco-workbench .part.auxiliarybar > .title > .pane-composite-bar > .monaco-action-bar .action-item:hover .action-label,
+.monaco-workbench .part.auxiliarybar > .title > .pane-composite-bar > .monaco-action-bar .action-item:focus .action-label {
 	color: var(--vscode-sideBarTitle-foreground) !important;
 }
 
-.monaco-workbench .part.auxiliarybar > .title > .panel-switcher-container > .monaco-action-bar .action-item.checked .action-label,
-.monaco-workbench .part.auxiliarybar > .title > .panel-switcher-container > .monaco-action-bar .action-item:hover .action-label {
+.monaco-workbench .part.auxiliarybar > .title > .pane-composite-bar > .monaco-action-bar .action-item.checked .action-label,
+.monaco-workbench .part.auxiliarybar > .title > .pane-composite-bar > .monaco-action-bar .action-item:hover .action-label {
 	outline: var(--vscode-contrastActiveBorder, unset) solid 1px !important;
 }
 
-.monaco-workbench .part.auxiliarybar > .title > .panel-switcher-container > .monaco-action-bar .action-item:not(.checked):hover .action-label {
+.monaco-workbench .part.auxiliarybar > .title > .pane-composite-bar > .monaco-action-bar .action-item:not(.checked):hover .action-label {
 	outline: var(--vscode-contrastActiveBorder, unset) dashed 1px !important;
 }
