+++ src/vs/workbench/browser/parts/paneCompositePart.ts
@@ -3,24 +3,44 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import 'vs/css!./media/paneCompositePart';
 import { Event } from 'vs/base/common/event';
-import { assertIsDefined } from 'vs/base/common/types';
-import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IProgressIndicator } from 'vs/platform/progress/common/progress';
-import { PaneCompositeDescriptor } from 'vs/workbench/browser/panecomposite';
-import { ActivitybarPart } from 'vs/workbench/browser/parts/activitybar/activitybarPart';
-import { AuxiliaryBarPart } from 'vs/workbench/browser/parts/auxiliarybar/auxiliaryBarPart';
-import { PanelPart } from 'vs/workbench/browser/parts/panel/panelPart';
-import { SidebarPart } from 'vs/workbench/browser/parts/sidebar/sidebarPart';
+import { Extensions, PaneComposite, PaneCompositeDescriptor, PaneCompositeRegistry } from 'vs/workbench/browser/panecomposite';
+// import { PanelPart } from 'vs/workbench/browser/parts/panel/panelPart';
 import { IPaneComposite } from 'vs/workbench/common/panecomposite';
-import { ViewContainerLocation, ViewContainerLocations } from 'vs/workbench/common/views';
+import { IViewDescriptorService, ViewContainerLocation } from 'vs/workbench/common/views';
 import { IBadge } from 'vs/workbench/services/activity/common/activity';
-import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
-import { Disposable, DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
+import { Disposable, DisposableStore, IDisposable, MutableDisposable } from 'vs/base/common/lifecycle';
 import { IView } from 'vs/base/browser/ui/grid/grid';
+import { IWorkbenchLayoutService, Parts } from 'vs/workbench/services/layout/browser/layoutService';
+import { CompositePart, ICompositeTitleLabel } from 'vs/workbench/browser/parts/compositePart';
+import { IPaneCompositeBarOptions, PaneCompositeBar } from 'vs/workbench/browser/parts/paneCompositeBar';
+import { Dimension, EventHelper, clearNode, trackFocus, $, addDisposableListener, EventType } from 'vs/base/browser/dom';
+import { Registry } from 'vs/platform/registry/common/platform';
+import { INotificationService } from 'vs/platform/notification/common/notification';
+import { IStorageService } from 'vs/platform/storage/common/storage';
+import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
+import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { IContextKey, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
+import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
+import { IComposite } from 'vs/workbench/common/composite';
+import { localize } from 'vs/nls';
+import { CompositeDragAndDropObserver, toggleDropEffect } from 'vs/workbench/browser/dnd';
+import { EDITOR_DRAG_AND_DROP_BACKGROUND } from 'vs/workbench/common/theme';
+import { IPartOptions } from 'vs/workbench/browser/part';
+import { ToolBar } from 'vs/base/browser/ui/toolbar/toolbar';
+import { CompositeMenuActions } from 'vs/workbench/browser/actions';
+import { MenuId } from 'vs/platform/actions/common/actions';
+import { ActionsOrientation, prepareActions } from 'vs/base/browser/ui/actionbar/actionbar';
+import { Gesture, EventType as GestureEventType } from 'vs/base/browser/touch';
+import { StandardMouseEvent } from 'vs/base/browser/mouseEvent';
 
 export interface IPaneCompositePart extends IView {
+
+	readonly partId: Parts.PANEL_PART | Parts.AUXILIARYBAR_PART | Parts.SIDEBAR_PART;
 
 	readonly onDidPaneCompositeOpen: Event<IPaneComposite>;
 	readonly onDidPaneCompositeClose: Event<IPaneComposite>;
@@ -79,84 +99,402 @@ export interface IPaneCompositeSelectorPart {
 	showActivity(id: string, badge: IBadge, clazz?: string, priority?: number): IDisposable;
 }
 
-export class PaneCompositeParts extends Disposable implements IPaneCompositePartService {
+export abstract class AbstractPaneCompositePart extends CompositePart<PaneComposite> implements IPaneCompositePart, IPaneCompositeSelectorPart {
 
-	declare readonly _serviceBrand: undefined;
+	private static readonly MIN_COMPOSITE_BAR_WIDTH = 50;
 
-	readonly onDidPaneCompositeOpen: Event<{ composite: IPaneComposite; viewContainerLocation: ViewContainerLocation }>;
-	readonly onDidPaneCompositeClose: Event<{ composite: IPaneComposite; viewContainerLocation: ViewContainerLocation }>;
-
-	private readonly paneCompositeParts = new Map<ViewContainerLocation, IPaneCompositePart>();
-	private readonly paneCompositeSelectorParts = new Map<ViewContainerLocation, IPaneCompositeSelectorPart>();
-
-	constructor(@IInstantiationService instantiationService: IInstantiationService) {
-		super();
-
-		const panelPart = instantiationService.createInstance(PanelPart);
-		const sideBarPart = instantiationService.createInstance(SidebarPart);
-		const auxiliaryBarPart = instantiationService.createInstance(AuxiliaryBarPart);
-		const activityBarPart = instantiationService.createInstance(ActivitybarPart, sideBarPart);
-
-		this.paneCompositeParts.set(ViewContainerLocation.Panel, panelPart);
-		this.paneCompositeParts.set(ViewContainerLocation.Sidebar, sideBarPart);
-		this.paneCompositeParts.set(ViewContainerLocation.AuxiliaryBar, auxiliaryBarPart);
-
-		this.paneCompositeSelectorParts.set(ViewContainerLocation.Panel, panelPart);
-		this.paneCompositeSelectorParts.set(ViewContainerLocation.Sidebar, activityBarPart);
-		this.paneCompositeSelectorParts.set(ViewContainerLocation.AuxiliaryBar, auxiliaryBarPart);
-
-		const eventDisposables = this._register(new DisposableStore());
-		this.onDidPaneCompositeOpen = Event.any(...ViewContainerLocations.map(loc => Event.map(this.paneCompositeParts.get(loc)!.onDidPaneCompositeOpen, composite => { return { composite, viewContainerLocation: loc }; }, eventDisposables)));
-		this.onDidPaneCompositeClose = Event.any(...ViewContainerLocations.map(loc => Event.map(this.paneCompositeParts.get(loc)!.onDidPaneCompositeClose, composite => { return { composite, viewContainerLocation: loc }; }, eventDisposables)));
+	get snap(): boolean {
+		// Always allow snapping closed
+		// Only allow dragging open if the panel contains view containers
+		return this.layoutService.isVisible(this.partId) || !!this.paneCompositeBar.value?.getVisiblePaneCompositeIds().length;
 	}
 
-	openPaneComposite(id: string | undefined, viewContainerLocation: ViewContainerLocation, focus?: boolean): Promise<IPaneComposite | undefined> {
-		return this.getPartByLocation(viewContainerLocation).openPaneComposite(id, focus);
+	get onDidPaneCompositeOpen(): Event<IPaneComposite> { return Event.map(this.onDidCompositeOpen.event, compositeEvent => <IPaneComposite>compositeEvent.composite); }
+	readonly onDidPaneCompositeClose = this.onDidCompositeClose.event as Event<IPaneComposite>;
+
+	private readonly location: ViewContainerLocation;
+	private titleDisposables = this._register(new DisposableStore());
+	private titleContainer: HTMLElement | undefined;
+	private paneTitleLabel: ICompositeTitleLabel | undefined;
+	private paneCompositeBar = this._register(new MutableDisposable<PaneCompositeBar>());
+	private emptyPaneMessageElement: HTMLElement | undefined;
+
+	private globalToolBar: ToolBar | undefined;
+	private readonly globalActions: CompositeMenuActions;
+
+	private blockOpening = false;
+	protected contentDimension: Dimension | undefined;
+
+	constructor(
+		readonly partId: Parts.PANEL_PART | Parts.AUXILIARYBAR_PART | Parts.SIDEBAR_PART,
+		partOptions: IPartOptions,
+		private readonly paneCompositeBarOptions: IPaneCompositeBarOptions,
+		activePaneCompositeSettingsKey: string,
+		private readonly activePaneContextKey: IContextKey<string>,
+		private paneFocusContextKey: IContextKey<boolean>,
+		nameForTelemetry: string,
+		compositeCSSClass: string,
+		titleForegroundColor: string | undefined,
+		@INotificationService notificationService: INotificationService,
+		@IStorageService storageService: IStorageService,
+		@IContextMenuService contextMenuService: IContextMenuService,
+		@IWorkbenchLayoutService layoutService: IWorkbenchLayoutService,
+		@IKeybindingService keybindingService: IKeybindingService,
+		@IInstantiationService instantiationService: IInstantiationService,
+		@IThemeService themeService: IThemeService,
+		@IViewDescriptorService private readonly viewDescriptorService: IViewDescriptorService,
+		@IContextKeyService protected readonly contextKeyService: IContextKeyService,
+		@IExtensionService private readonly extensionService: IExtensionService,
+	) {
+		let location = ViewContainerLocation.Sidebar;
+		let registryId = Extensions.Viewlets;
+		let globalActionsMenuId = MenuId.SidebarTitle;
+		if (partId === Parts.PANEL_PART) {
+			location = ViewContainerLocation.Panel;
+			registryId = Extensions.Panels;
+			globalActionsMenuId = MenuId.PanelTitle;
+		} else if (partId === Parts.AUXILIARYBAR_PART) {
+			location = ViewContainerLocation.AuxiliaryBar;
+			registryId = Extensions.Auxiliary;
+			globalActionsMenuId = MenuId.AuxiliaryBarTitle;
+		}
+		super(
+			notificationService,
+			storageService,
+			contextMenuService,
+			layoutService,
+			keybindingService,
+			instantiationService,
+			themeService,
+			Registry.as<PaneCompositeRegistry>(registryId),
+			activePaneCompositeSettingsKey,
+			viewDescriptorService.getDefaultViewContainer(location)?.id || '',
+			nameForTelemetry,
+			compositeCSSClass,
+			titleForegroundColor,
+			partId,
+			partOptions
+		);
+
+		this.location = location;
+		this.globalActions = this._register(this.instantiationService.createInstance(CompositeMenuActions, globalActionsMenuId, undefined, undefined));
+
+		this.registerListeners();
 	}
 
-	getActivePaneComposite(viewContainerLocation: ViewContainerLocation): IPaneComposite | undefined {
-		return this.getPartByLocation(viewContainerLocation).getActivePaneComposite();
+	private registerListeners(): void {
+		this._register(this.onDidPaneCompositeOpen(composite => this.onDidOpen(composite)));
+		this._register(this.onDidPaneCompositeClose(this.onDidClose, this));
+		this._register(this.globalActions.onDidChange(() => this.updateGlobalToolbarActions()));
+
+		// TODO: @sandy081 Is this needed?
+		this._register(this.registry.onDidDeregister(async (viewletDescriptor: PaneCompositeDescriptor) => {
+
+			const activeContainers = this.viewDescriptorService.getViewContainersByLocation(this.location)
+				.filter(container => this.viewDescriptorService.getViewContainerModel(container).activeViewDescriptors.length > 0);
+
+			if (activeContainers.length) {
+				if (this.getActiveComposite()?.getId() === viewletDescriptor.id) {
+					const defaultViewletId = this.viewDescriptorService.getDefaultViewContainer(this.location)?.id;
+					const containerToOpen = activeContainers.filter(c => c.id === defaultViewletId)[0] || activeContainers[0];
+					await this.openPaneComposite(containerToOpen.id);
+				}
+			} else {
+				this.layoutService.setPartHidden(true, this.partId);
+			}
+
+			this.removeComposite(viewletDescriptor.id);
+		}));
 	}
 
-	getPaneComposite(id: string, viewContainerLocation: ViewContainerLocation): PaneCompositeDescriptor | undefined {
-		return this.getPartByLocation(viewContainerLocation).getPaneComposite(id);
+	private onDidOpen(composite: IComposite): void {
+		this.activePaneContextKey.set(composite.getId());
+		this.layoutEmptyMessage();
 	}
 
-	getPaneComposites(viewContainerLocation: ViewContainerLocation): PaneCompositeDescriptor[] {
-		return this.getPartByLocation(viewContainerLocation).getPaneComposites();
+	private onDidClose(composite: IComposite): void {
+		const id = composite.getId();
+		if (this.activePaneContextKey.get() === id) {
+			this.activePaneContextKey.reset();
+		}
+		this.layoutEmptyMessage();
 	}
 
-	getPinnedPaneCompositeIds(viewContainerLocation: ViewContainerLocation): string[] {
-		return this.getSelectorPartByLocation(viewContainerLocation).getPinnedPaneCompositeIds();
+	override create(parent: HTMLElement): void {
+		this.element = parent;
+
+		super.create(parent);
+
+		this.createEmptyPaneMessage();
+
+		const focusTracker = this._register(trackFocus(parent));
+		this._register(focusTracker.onDidFocus(() => this.paneFocusContextKey.set(true)));
+		this._register(focusTracker.onDidBlur(() => this.paneFocusContextKey.set(false)));
 	}
 
-	getVisiblePaneCompositeIds(viewContainerLocation: ViewContainerLocation): string[] {
-		return this.getSelectorPartByLocation(viewContainerLocation).getVisiblePaneCompositeIds();
+	private createEmptyPaneMessage(): void {
+		const contentArea = this.getContentArea()!;
+		this.emptyPaneMessageElement = document.createElement('div');
+		this.emptyPaneMessageElement.classList.add('empty-pane-message-area');
+
+		const messageElement = document.createElement('div');
+		messageElement.classList.add('empty-pane-message');
+		messageElement.innerText = localize('pane.emptyMessage', "Drag a view here to display.");
+
+		this.emptyPaneMessageElement.appendChild(messageElement);
+		contentArea.appendChild(this.emptyPaneMessageElement);
+
+		this._register(CompositeDragAndDropObserver.INSTANCE.registerTarget(this.emptyPaneMessageElement, {
+			onDragOver: (e) => {
+				EventHelper.stop(e.eventData, true);
+				if (this.paneCompositeBar.value) {
+					const validDropTarget = this.paneCompositeBar.value.dndHandler.onDragEnter(e.dragAndDropData, undefined, e.eventData);
+					toggleDropEffect(e.eventData.dataTransfer, 'move', validDropTarget);
+				}
+			},
+			onDragEnter: (e) => {
+				EventHelper.stop(e.eventData, true);
+				if (this.paneCompositeBar.value) {
+					const validDropTarget = this.paneCompositeBar.value.dndHandler.onDragEnter(e.dragAndDropData, undefined, e.eventData);
+					this.emptyPaneMessageElement!.style.backgroundColor = validDropTarget ? this.theme.getColor(EDITOR_DRAG_AND_DROP_BACKGROUND)?.toString() || '' : '';
+				}
+			},
+			onDragLeave: (e) => {
+				EventHelper.stop(e.eventData, true);
+				this.emptyPaneMessageElement!.style.backgroundColor = '';
+			},
+			onDragEnd: (e) => {
+				EventHelper.stop(e.eventData, true);
+				this.emptyPaneMessageElement!.style.backgroundColor = '';
+			},
+			onDrop: (e) => {
+				EventHelper.stop(e.eventData, true);
+				this.emptyPaneMessageElement!.style.backgroundColor = '';
+				if (this.paneCompositeBar.value) {
+					this.paneCompositeBar.value.dndHandler.drop(e.dragAndDropData, undefined, e.eventData);
+				}
+			},
+		}));
 	}
 
-	getProgressIndicator(id: string, viewContainerLocation: ViewContainerLocation): IProgressIndicator | undefined {
-		return this.getPartByLocation(viewContainerLocation).getProgressIndicator(id);
+	async openPaneComposite(id?: string, focus?: boolean): Promise<PaneComposite | undefined> {
+		if (typeof id === 'string' && this.getPaneComposite(id)) {
+			return this.doOpenPaneComposite(id, focus);
+		}
+
+		await this.extensionService.whenInstalledExtensionsRegistered();
+
+		if (typeof id === 'string' && this.getPaneComposite(id)) {
+			return this.doOpenPaneComposite(id, focus);
+		}
+
+		return undefined;
 	}
 
-	hideActivePaneComposite(viewContainerLocation: ViewContainerLocation): void {
-		this.getPartByLocation(viewContainerLocation).hideActivePaneComposite();
+	private doOpenPaneComposite(id: string, focus?: boolean): PaneComposite | undefined {
+		if (this.blockOpening) {
+			return undefined; // Workaround against a potential race condition
+		}
+
+		if (!this.layoutService.isVisible(this.partId)) {
+			try {
+				this.blockOpening = true;
+				this.layoutService.setPartHidden(false, this.partId);
+			} finally {
+				this.blockOpening = false;
+			}
+		}
+
+		return this.openComposite(id, focus) as PaneComposite;
 	}
 
-	getLastActivePaneCompositeId(viewContainerLocation: ViewContainerLocation): string {
-		return this.getPartByLocation(viewContainerLocation).getLastActivePaneCompositeId();
+	showActivity(id: string, badge: IBadge, clazz?: string, priority?: number): IDisposable {
+		return this.paneCompositeBar.value?.showActivity(id, badge, clazz, priority) ?? Disposable.None;
 	}
 
-	showActivity(id: string, viewContainerLocation: ViewContainerLocation, badge: IBadge, clazz?: string, priority?: number): IDisposable {
-		return this.getSelectorPartByLocation(viewContainerLocation).showActivity(id, badge, clazz, priority);
+	getPaneComposite(id: string): PaneCompositeDescriptor | undefined {
+		return (this.registry as PaneCompositeRegistry).getPaneComposite(id);
 	}
 
-	private getPartByLocation(viewContainerLocation: ViewContainerLocation): IPaneCompositePart {
-		return assertIsDefined(this.paneCompositeParts.get(viewContainerLocation));
+	getPaneComposites(): PaneCompositeDescriptor[] {
+		return (this.registry as PaneCompositeRegistry).getPaneComposites()
+			.sort((v1, v2) => {
+				if (typeof v1.order !== 'number') {
+					return 1;
+				}
+
+				if (typeof v2.order !== 'number') {
+					return -1;
+				}
+
+				return v1.order - v2.order;
+			});
 	}
 
-	private getSelectorPartByLocation(viewContainerLocation: ViewContainerLocation): IPaneCompositeSelectorPart {
-		return assertIsDefined(this.paneCompositeSelectorParts.get(viewContainerLocation));
+	getPinnedPaneCompositeIds(): string[] {
+		return this.paneCompositeBar.value?.getPinnedPaneCompositeIds() ?? [];
 	}
+
+	getVisiblePaneCompositeIds(): string[] {
+		return this.paneCompositeBar.value?.getVisiblePaneCompositeIds() ?? [];
+	}
+
+	getActivePaneComposite(): IPaneComposite | undefined {
+		return <IPaneComposite>this.getActiveComposite();
+	}
+
+	getLastActivePaneCompositeId(): string {
+		return this.getLastActiveCompositeId();
+	}
+
+	hideActivePaneComposite(): void {
+		if (this.layoutService.isVisible(this.partId)) {
+			this.layoutService.setPartHidden(true, this.partId);
+		}
+
+		this.hideActiveComposite();
+	}
+
+	protected override createTitleArea(parent: HTMLElement): HTMLElement {
+		const element = super.createTitleArea(parent);
+
+		this._register(addDisposableListener(element, EventType.CONTEXT_MENU, e => {
+			this.onTitleAreaContextMenu(new StandardMouseEvent(e));
+		}));
+		this._register(Gesture.addTarget(element));
+		this._register(addDisposableListener(element, GestureEventType.Contextmenu, e => {
+			this.onTitleAreaContextMenu(new StandardMouseEvent(e));
+		}));
+
+		const globalTitleActionsContainer = element.appendChild($('.global-actions'));
+
+		// Global Actions Toolbar
+		this.globalToolBar = this._register(new ToolBar(globalTitleActionsContainer, this.contextMenuService, {
+			actionViewItemProvider: action => this.actionViewItemProvider(action),
+			orientation: ActionsOrientation.HORIZONTAL,
+			getKeyBinding: action => this.keybindingService.lookupKeybinding(action.id),
+			anchorAlignmentProvider: () => this.getTitleAreaDropDownAnchorAlignment(),
+			toggleMenuTitle: localize('moreActions', "More Actions...")
+		}));
+
+		this.updateGlobalToolbarActions();
+
+		return element;
+	}
+
+	protected override createTitleLabel(parent: HTMLElement): ICompositeTitleLabel {
+		this.titleContainer = parent;
+		this.updateTitleArea();
+		return {
+			updateTitle: (id, title, keybinding) => {
+				if (!this.updateTitleArea() && this.paneTitleLabel) {
+					this.paneTitleLabel.updateTitle(id, title, keybinding);
+				}
+			},
+			updateStyles: () => this.paneTitleLabel?.updateStyles()
+		};
+	}
+
+	protected updateTitleArea(): boolean {
+		if (!this.titleContainer) {
+			return false;
+		}
+		if (!this.paneCompositeBar.value && this.shouldShowCompositeBar()) {
+			this.titleDisposables.clear();
+			this.titleLabelElement = undefined;
+			clearNode(this.titleContainer);
+			this.paneCompositeBar.value = this.instantiationService.createInstance(PaneCompositeBar, this.paneCompositeBarOptions, this.partId, this);
+			const titleArea = this.paneCompositeBar.value.create(this.titleContainer);
+			titleArea.classList.add('pane-composite-bar');
+			return true;
+		}
+		if (!this.titleLabelElement && !this.shouldShowCompositeBar()) {
+			this.paneCompositeBar.clear();
+			this.titleDisposables.clear();
+			clearNode(this.titleContainer);
+			this.paneTitleLabel = super.createTitleLabel(this.titleContainer);
+			this.titleLabelElement!.draggable = true;
+			const draggedItemProvider = (): { type: 'view' | 'composite'; id: string } => {
+				const activeViewlet = this.getActivePaneComposite()!;
+				return { type: 'composite', id: activeViewlet.getId() };
+			};
+			this.titleDisposables.add(CompositeDragAndDropObserver.INSTANCE.registerDraggable(this.titleLabelElement!, draggedItemProvider, {}));
+			return false;
+		}
+		return false;
+	}
+
+	protected override onTitleAreaUpdate(compositeId: string): void {
+		super.onTitleAreaUpdate(compositeId);
+
+		// If title actions change, relayout the composite bar
+		this.layoutCompositeBar();
+	}
+
+	override layout(width: number, height: number, top: number, left: number): void {
+		if (!this.layoutService.isVisible(this.partId)) {
+			return;
+		}
+
+		this.contentDimension = new Dimension(width, height);
+
+		// Layout contents
+		super.layout(this.contentDimension.width, this.contentDimension.height, top, left);
+
+		// Layout composite bar
+		this.layoutCompositeBar();
+
+		// Add empty pane message
+		this.layoutEmptyMessage();
+	}
+
+	private layoutCompositeBar(): void {
+		if (this.contentDimension && this.dimension && this.paneCompositeBar.value) {
+			let availableWidth = this.contentDimension.width - 40; // take padding into account
+			if (this.toolBar) {
+				availableWidth = Math.max(AbstractPaneCompositePart.MIN_COMPOSITE_BAR_WIDTH, availableWidth - this.getToolbarWidth());
+			}
+			this.paneCompositeBar.value.layout(availableWidth, this.dimension.height);
+		}
+	}
+
+	private layoutEmptyMessage(): void {
+		this.emptyPaneMessageElement?.classList.toggle('visible', !!this.paneCompositeBar.value && this.paneCompositeBar.value.getVisiblePaneCompositeIds().length === 0);
+	}
+
+	private updateGlobalToolbarActions(): void {
+		const primaryActions = this.globalActions.getPrimaryActions();
+		const secondaryActions = this.globalActions.getSecondaryActions();
+		this.globalToolBar?.setActions(prepareActions(primaryActions), prepareActions(secondaryActions));
+	}
+
+	protected getToolbarWidth(): number {
+		const activePane = this.getActivePaneComposite();
+		if (!activePane || !this.toolBar) {
+			return 0;
+		}
+		return this.toolBar.getItemsWidth() + (this.globalToolBar?.getItemsWidth() ?? 0);
+	}
+
+	private onTitleAreaContextMenu(event: StandardMouseEvent): void {
+		if (this.shouldShowCompositeBar()) {
+			return;
+		}
+		const activeViewlet = this.getActivePaneComposite() as PaneComposite;
+		if (activeViewlet) {
+			const contextMenuActions = activeViewlet ? activeViewlet.getContextMenuActions() : [];
+			if (contextMenuActions.length) {
+				this.contextMenuService.showContextMenu({
+					getAnchor: () => event,
+					getActions: () => contextMenuActions.slice(),
+					getActionViewItem: action => this.actionViewItemProvider(action),
+					actionRunner: activeViewlet.getActionRunner(),
+					skipTelemetry: true
+				});
+			}
+		}
+	}
+
+	protected abstract shouldShowCompositeBar(): boolean;
+
 }
-
-registerSingleton(IPaneCompositePartService, PaneCompositeParts, InstantiationType.Delayed);
