+++ src/vs/workbench/browser/parts/paneCompositePartService.ts
@@ -1,0 +1,103 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Event } from 'vs/base/common/event';
+import { assertIsDefined } from 'vs/base/common/types';
+import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { IProgressIndicator } from 'vs/platform/progress/common/progress';
+import { PaneCompositeDescriptor } from 'vs/workbench/browser/panecomposite';
+import { AuxiliaryBarPart } from 'vs/workbench/browser/parts/auxiliarybar/auxiliaryBarPart';
+import { PanelPart } from 'vs/workbench/browser/parts/panel/panelPart';
+import { SidebarPart } from 'vs/workbench/browser/parts/sidebar/sidebarPart';
+import { IPaneComposite } from 'vs/workbench/common/panecomposite';
+import { ViewContainerLocation, ViewContainerLocations } from 'vs/workbench/common/views';
+import { IBadge } from 'vs/workbench/services/activity/common/activity';
+import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
+import { Disposable, DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
+import { IPaneCompositePart, IPaneCompositeSelectorPart } from 'vs/workbench/browser/parts/paneCompositePart';
+
+export class PaneCompositePartService extends Disposable implements IPaneCompositePartService {
+
+	declare readonly _serviceBrand: undefined;
+
+	readonly onDidPaneCompositeOpen: Event<{ composite: IPaneComposite; viewContainerLocation: ViewContainerLocation }>;
+	readonly onDidPaneCompositeClose: Event<{ composite: IPaneComposite; viewContainerLocation: ViewContainerLocation }>;
+
+	private readonly paneCompositeParts = new Map<ViewContainerLocation, IPaneCompositePart>();
+	private readonly paneCompositeSelectorParts = new Map<ViewContainerLocation, IPaneCompositeSelectorPart>();
+
+	constructor(
+		@IInstantiationService instantiationService: IInstantiationService,
+	) {
+		super();
+
+		const panelPart = instantiationService.createInstance(PanelPart);
+		const sideBarPart = instantiationService.createInstance(SidebarPart);
+		const auxiliaryBarPart = instantiationService.createInstance(AuxiliaryBarPart);
+
+		this.paneCompositeParts.set(ViewContainerLocation.Panel, panelPart);
+		this.paneCompositeParts.set(ViewContainerLocation.Sidebar, sideBarPart);
+		this.paneCompositeParts.set(ViewContainerLocation.AuxiliaryBar, auxiliaryBarPart);
+
+		this.paneCompositeSelectorParts.set(ViewContainerLocation.Panel, panelPart);
+		this.paneCompositeSelectorParts.set(ViewContainerLocation.Sidebar, sideBarPart);
+		this.paneCompositeSelectorParts.set(ViewContainerLocation.AuxiliaryBar, auxiliaryBarPart);
+
+		const eventDisposables = this._register(new DisposableStore());
+		this.onDidPaneCompositeOpen = Event.any(...ViewContainerLocations.map(loc => Event.map(this.paneCompositeParts.get(loc)!.onDidPaneCompositeOpen, composite => { return { composite, viewContainerLocation: loc }; }, eventDisposables)));
+		this.onDidPaneCompositeClose = Event.any(...ViewContainerLocations.map(loc => Event.map(this.paneCompositeParts.get(loc)!.onDidPaneCompositeClose, composite => { return { composite, viewContainerLocation: loc }; }, eventDisposables)));
+	}
+
+	openPaneComposite(id: string | undefined, viewContainerLocation: ViewContainerLocation, focus?: boolean): Promise<IPaneComposite | undefined> {
+		return this.getPartByLocation(viewContainerLocation).openPaneComposite(id, focus);
+	}
+
+	getActivePaneComposite(viewContainerLocation: ViewContainerLocation): IPaneComposite | undefined {
+		return this.getPartByLocation(viewContainerLocation).getActivePaneComposite();
+	}
+
+	getPaneComposite(id: string, viewContainerLocation: ViewContainerLocation): PaneCompositeDescriptor | undefined {
+		return this.getPartByLocation(viewContainerLocation).getPaneComposite(id);
+	}
+
+	getPaneComposites(viewContainerLocation: ViewContainerLocation): PaneCompositeDescriptor[] {
+		return this.getPartByLocation(viewContainerLocation).getPaneComposites();
+	}
+
+	getPinnedPaneCompositeIds(viewContainerLocation: ViewContainerLocation): string[] {
+		return this.getSelectorPartByLocation(viewContainerLocation).getPinnedPaneCompositeIds();
+	}
+
+	getVisiblePaneCompositeIds(viewContainerLocation: ViewContainerLocation): string[] {
+		return this.getSelectorPartByLocation(viewContainerLocation).getVisiblePaneCompositeIds();
+	}
+
+	getProgressIndicator(id: string, viewContainerLocation: ViewContainerLocation): IProgressIndicator | undefined {
+		return this.getPartByLocation(viewContainerLocation).getProgressIndicator(id);
+	}
+
+	hideActivePaneComposite(viewContainerLocation: ViewContainerLocation): void {
+		this.getPartByLocation(viewContainerLocation).hideActivePaneComposite();
+	}
+
+	getLastActivePaneCompositeId(viewContainerLocation: ViewContainerLocation): string {
+		return this.getPartByLocation(viewContainerLocation).getLastActivePaneCompositeId();
+	}
+
+	showActivity(id: string, viewContainerLocation: ViewContainerLocation, badge: IBadge, clazz?: string, priority?: number): IDisposable {
+		return this.getSelectorPartByLocation(viewContainerLocation).showActivity(id, badge, clazz, priority);
+	}
+
+	private getPartByLocation(viewContainerLocation: ViewContainerLocation): IPaneCompositePart {
+		return assertIsDefined(this.paneCompositeParts.get(viewContainerLocation));
+	}
+
+	private getSelectorPartByLocation(viewContainerLocation: ViewContainerLocation): IPaneCompositeSelectorPart {
+		return assertIsDefined(this.paneCompositeSelectorParts.get(viewContainerLocation));
+	}
+}
+
+registerSingleton(IPaneCompositePartService, PaneCompositePartService, InstantiationType.Delayed);
