+++ src/vs/workbench/workbench.common.main.ts
@@ -43,7 +43,7 @@ import 'vs/workbench/api/browser/viewsExtensionPoint';
 
 import 'vs/workbench/browser/parts/editor/editor.contribution';
 import 'vs/workbench/browser/parts/editor/editorParts';
-import 'vs/workbench/browser/parts/paneCompositePart';
+import 'vs/workbench/browser/parts/paneCompositePartService';
 import 'vs/workbench/browser/parts/banner/bannerPart';
 import 'vs/workbench/browser/parts/statusbar/statusbarPart';
 import 'vs/workbench/browser/parts/views/viewsService';
